backend/requirements.txt
Flask==2.2.2
pymongo==4.3.3
python-dotenv==0.21.0
Flask-Cors==3.0.10
Werkzeug==2.3.8
certifi==2024.7.4